Final refactor on the whole system by all

+++ springboot-first-app/src/main/java/app/filerepository/delete/controller/Delete.java
@@ -1,6 +1,7 @@
 package app.filerepository.delete.controller;
 
-import app.config.TokenProvider;
+import app.exception.DeleteException;
+import app.userauth.config.TokenProvider;
 import app.response.ResponseMessage;
 import app.filerepository.delete.service.intf.IDeleteService;
 import org.slf4j.Logger;
@@ -11,7 +12,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
-
+import static app.filerepository.delete.constant.DeleteEndPoint.*;
+import static app.filerepository.delete.constant.ResponseConstant.*;
 import static app.userauth.role.RoleEnum.ADMIN;
 
 @RestController
@@ -19,22 +21,26 @@ public class Delete {
     @Autowired
     private IDeleteService deleteService;
     @Autowired
-    TokenProvider jwtTokenUtils;
+    private TokenProvider jwtTokenUtils;
     private static final Logger logger = LoggerFactory.getLogger(Delete.class);
 
-    @DeleteMapping("/delete/{fileId}")
-    public ResponseMessage deleteFile(@PathVariable String fileId){
-        logger.info("create delete response for "+fileId);
-        return deleteService.delete(fileId);
+    @DeleteMapping(DELETE_BY_ID_END_POINT)
+    public ResponseMessage deleteFile(@PathVariable String fileId, HttpServletRequest request) throws DeleteException {
+        String role = jwtTokenUtils.getRole(request);
+        if(role.equalsIgnoreCase(ADMIN.name())){
+            logger.info("create delete response for "+fileId);
+            return deleteService.delete(fileId);
+        }
+        return ResponseMessage.getInstance(NO_PERMISSION_TO_DELETE_FILE);
     }
 
-    @DeleteMapping("/deleteAll")
-    public ResponseMessage deleteAllFiles(HttpServletRequest request){
-        String role =   jwtTokenUtils.getRole(request);
+    @DeleteMapping(DELETE_ALL_FILE_END_POINT)
+    public ResponseMessage deleteAllFiles(HttpServletRequest request) throws DeleteException {
+        String role = jwtTokenUtils.getRole(request);
         logger.info("deleteAllFiles Response created ");
         if(role.equalsIgnoreCase(ADMIN.name())){
             return deleteService.deleteAllFiles();
         }
-        return ResponseMessage.getInstance("you don't have permission to delete files");
+        return ResponseMessage.getInstance(NO_PERMISSION_TO_DELETE_FILE);
     }
 }
